agent-blog-post/coding-agent/pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
